+++ laktory/models/compute/__init__.py
@@ -3,3 +3,4 @@ from laktory.models.compute.cluster import Cluster
 from laktory.models.compute.warehouse import Warehouse
 from laktory.models.compute.notebook import Notebook
 from laktory.models.compute.pipeline import Pipeline
+from laktory.models.compute.job import Job